1133-dev-overarching-timeseries-client: Remove public modifier for all methods in RDB client and Sparql client which modify time series to limit visibility to package scope and limit use to methods to overarching time series client; move all tests to package timeseries (from timeseries/test)
